requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
blessed==1.17.6
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coloredlogs==15.0
dataclasses-json==0.5.3
EasyProcess==0.3
entrypoint2==0.2.4
file-downloader==0.5.1
filetype==1.0.7
fuzzywuzzy==0.18.0
gitdb==4.0.7
GitPython==3.1.14
giturlparse.py==0.0.5
humanfriendly==9.1
idna==2.10
iniconfig==1.1.1
inquirer==2.7.0
libarchive==0.4.7
linecache2==1.0.0
marshmallow==3.11.1
marshmallow-enum==1.5.1
more-itertools==8.7.0
mypy==0.812
mypy-extensions==0.4.3
nose==1.3.7
packaging==20.9
patool==1.12
pep8==1.7.1
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-Levenshtein==0.12.2
pytz==2021.1
pyunpack==0.2.2
PyYAML==5.4.1
readchar==2.0.1
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.16.0
smmap==4.0.0
stringcase==1.2.0
tabulate==0.8.9
toml==0.10.2
tqdm==4.60.0
traceback2==1.4.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
typing-inspect==0.6.0
unittest2==1.1.0
urllib3==1.26.4
wcwidth==0.2.5
